230711
최종 백업

+++ 프레임워크 기본틀/230608/Attack_Bar.h
@@ -19,6 +19,6 @@ private:
 	int		Damage_size(float _fX);
 
 private:
-	DWORD		m_dwTime = GetTickCount64();
+	DWORD		m_dwTime = (DWORD)GetTickCount64();
 };
