+++ 프레임워크 기본틀/230608/Attack_Motion.h
@@ -15,7 +15,7 @@ public:
 	virtual void Release(void)		override;
 
 private:
-	DWORD		m_dwTime = GetTickCount64();
+	DWORD		m_dwTime = (DWORD)GetTickCount64();
 	TCHAR	szBuff[100] = L"";
 
 };
