+++ 프레임워크 기본틀/230608/LookBullet.cpp
@@ -4,7 +4,7 @@
 #include "SceneMgr.h"
 
 CLooKBullet::CLooKBullet()
-	:m_dwTime(GetTickCount64()), m_iTime(0)
+	:m_dwTime((DWORD)GetTickCount64()), m_iTime(0)
 {
 }
