+++ 프레임워크 기본틀/230608/LookNormalBullet.cpp
@@ -6,7 +6,7 @@
 #include "SceneMgr.h"
 
 CLooKNormalBullet::CLooKNormalBullet()
-	:m_dwTime(GetTickCount64()), m_iTime(0)
+	:m_dwTime((DWORD)GetTickCount64()), m_iTime(0)
 {
 }
 
@@ -22,7 +22,7 @@ void CLooKNormalBullet::Initialize(void)
 
 	m_fSpeed = 2.f;
 	m_fAngle = 5.f;
-	m_iAttack = 2.f;
+	m_iAttack = 2;
 
 	CBmpMgr::Get_Instance()->Insert_Bmp(L"../Resource/Monster/Look_Attack.bmp", L"Look_Attack");
 
@@ -43,9 +43,9 @@ int CLooKNormalBullet::Update(void)
 	m_tInfo.fX += m_fSpeed * cosf(m_fAngle * (PI / 180.f));
 	m_tInfo.fY -= m_fSpeed * sinf(m_fAngle * (PI / 180.f));
 
-	if (m_dwTime + 400.f < GetTickCount64()) {
+	if (m_dwTime + 400.f < (DWORD)GetTickCount64()) {
 		CObjMgr::Get_Instance()->Add_Object(BULLET, Create_Bullet());
-		m_dwTime = GetTickCount64();
+		m_dwTime = (DWORD)GetTickCount64();
 	}
 	__super::Update_Rect();
