+++ 프레임워크 기본틀/230608/MainGame.cpp
@@ -8,7 +8,7 @@
 #include "SoundMgr.h"
 #include "BmpMgr.h"
 
-CMainGame::CMainGame() : m_dwTime(GetTickCount64()), m_iFPS(0)
+CMainGame::CMainGame() : m_dwTime((DWORD)GetTickCount64()), m_iFPS(0)
 {
 	ZeroMemory(m_szFPS, sizeof(m_szFPS));
 }
@@ -29,7 +29,6 @@ void CMainGame::Initialize()
 
 	CSceneMgr::Get_Instance()->Scene_Change(INTRO);
 
-
 }
 
 void CMainGame::Update()
@@ -47,13 +46,13 @@ void CMainGame::Render()
 {
 	++m_iFPS;
 
-	if (m_dwTime + 2000 < GetTickCount64())
+	if (m_dwTime + 2000 < (DWORD)GetTickCount64())
 	{
 		swprintf_s(m_szFPS, L"FPS : %d", m_iFPS);
 		SetWindowText(g_hWnd, m_szFPS);
 
 		m_iFPS = 0;
-		m_dwTime = GetTickCount64();
+		m_dwTime = (DWORD)GetTickCount64();
 	}
 
 	AddFontResourceA("NeoµÕ±Ù¸ð.ttf");
@@ -71,7 +70,6 @@ void CMainGame::Render()
 	CSceneMgr::Get_Instance()->Render(hMemDC);
 
 	BitBlt(m_hDC, 0, 0, WINCX, WINCY, hMemDC, 0, 0, SRCCOPY);
-
 
 }
