+++ 프레임워크 기본틀/230608/MonsterMercy.h
@@ -18,7 +18,7 @@ public:
 	virtual void Release(void) override;
 
 private:
-	DWORD		m_preDeley = GetTickCount64();
-	DWORD		m_PostDelay = GetTickCount64();
+	DWORD		m_preDeley = (DWORD)GetTickCount64();
+	DWORD		m_PostDelay = (DWORD)GetTickCount64();
 };
