+++ 프레임워크 기본틀/230608/Stage_Ruin.cpp
@@ -57,7 +57,7 @@ void CStage1::Late_Update()
 
 	if (m_dwTime + 5000 < GetTickCount64()) {
 		CSceneMgr::Get_Instance()->Scene_Change(MONSTER_MATCH);
-		m_dwTime = GetTickCount64();
+		m_dwTime = (DWORD)GetTickCount64();
 		return;
 	}
 }
