+++ 프레임워크 기본틀/230608/Toriel.h
@@ -17,7 +17,7 @@ public:
 	virtual void		Release(void) override;
 
 private:
-	DWORD				m_dwTime = GetTickCount64();
+	DWORD				m_dwTime = (DWORD)GetTickCount64();
 	int					m_iDrawID;
 
 	TCHAR				szBuff[100] = L"";
